models/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0

# Machine learning utilities
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Optional: Weights & Biases logging
# wandb>=0.15.0

# Optional: Mixed precision training (included in torch)
# apex  # NVIDIA Apex (alternative to torch.cuda.amp)

# Optional: Age synthesis models (install separately)
# For SAM: https://github.com/yuval-alaluf/SAM
# For HRFAE: https://github.com/InterDigitalInc/HRFAE

# Development
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
